remove function removeUnderlying in Voter to avoid breaking update of vote
After internal discussion, we have decided to remove this function `removeUnderlying`; since the minter only reference the latest list of "reserve" asset for emission from the pool contract. A market would not be counted anyway if it is delisted from the lending pool. So user can still vote/devote on this pool but they just wont have any material impacts on the emission of $KZA.